+++ cypress/e2e/EnrollmentAddEventPage/sharedSteps.js
@@ -6,6 +6,8 @@ Then(/^you see the following (.*)$/, (message) => {
 
 And(/^you see the widget header (.*)$/, (name) => {
     cy.get('[data-test="add-event-enrollment-page-content"]')
+        .parent() // widget-contents
+        .parent() // widget container (WidgetNonCollapsible)
         .within(() => {
             cy.get('[data-test="widget-contents"]').should('exist');
             cy.get('[data-test="widget-header"]').should('exist');