+++ src/core_modules/capture-core/components/D2Form/field/configs/optionSetField/getOptionSetFieldConfig.ts
@@ -47,7 +47,6 @@ export const getOptionSetFieldConfig = (metaData: DataElement, options: any, que
         formHorizontal: options.formHorizontal,
         fieldLabelMediaBasedClass: options.fieldLabelMediaBasedClass,
         optionGroups: optionSet.optionGroups,
-        showHelpText: options.showHelpText,
         ...inputTypeProps,
     }, options, metaData);