+++ src/core_modules/capture-core/components/FiltersForTypes/EmptyValue/types/emptyValue.types.ts
@@ -2,6 +2,7 @@ export type EmptyValueFilterCheckboxesProps = {
     value?: string | null;
     onEmptyChange: (args: { checked: boolean }) => void;
     onNotEmptyChange: (args: { checked: boolean }) => void;
+    showDivider?: boolean;
 };
 
 export type EmptyValueFilterData = {