+++ src/core_modules/capture-core/components/FiltersForTypes/TrueOnly/TrueOnly.types.ts
@@ -3,5 +3,6 @@ export type Value = Array<any> | null | undefined;
 export type PlainProps = {
     value: Value;
     onCommitValue: (value: Value) => void;
+    onUpdate?: (value?: Value) => void;
 };