+++ src/core_modules/capture-core/components/FormFields/New/Fields/OrgUnitField/SingleOrgUnitSelectField.component.tsx
@@ -24,7 +24,7 @@ type OrgUnitValue = {
 };
 
 type SingleOrgUnitSelectFieldState = {
-    previousOrgUnitId: any;
+    previousOrgUnitId: string | null;
 };
 
 type SingleOrgUnitSelectFieldProps = {
@@ -52,7 +52,7 @@ class SingleOrgUnitSelectFieldPlain extends React.Component<Props, SingleOrgUnit
     }
 
     onDeselectOrgUnit = () => {
-        this.props.value && this.setState({ previousOrgUnitId: { id: this.props.value.id } });
+        this.props.value && this.setState({ previousOrgUnitId: this.props.value.id });
         this.props.onBlur(null);
     }