+++ src/core_modules/capture-core/components/ListView/Filters/FilterButton/buttonTextBuilder/converters/index.ts
@@ -1,6 +1,8 @@
 export { convertDate } from './dateConverter';
+export { convertTime } from './timeConverter';
 export { convertAssignee } from './assigneeConverter';
 export { convertBoolean } from './booleanConverter';
 export { convertText } from './textConverter';
 export { convertNumeric } from './numericConverter';
 export { convertTrueOnly } from './trueOnlyConverter';
+export { convertOrgUnit } from './orgUnitConverter';