+++ src/core_modules/capture-core/utils/validation/validators/form/isValidMinCharactersToSearch.ts
@@ -15,7 +15,6 @@ const isValueBiggerThanMinCharactersToSearch = (value: string, minCharactersToSe
     return minCharactersToSearch <= trimmedValue.length;
 };
 
-// eslint-disable-next-line complexity
 const isValidMinCharactersToSearchRange = (value: { from: any; to: any }, minCharactersToSearch: number) => {
     const { from, to } = value;
     if (from === undefined && to === undefined) {
@@ -32,17 +31,19 @@ const isValidMinCharactersToSearchRange = (value: { from: any; to: any }, minCha
 
     if (date) {
         return (
-            isValueBiggerThanMinCharactersToSearch(from.date, minCharactersToSearch) &&
-            isValueBiggerThanMinCharactersToSearch(to.date, minCharactersToSearch)
+            isValueBiggerThanMinCharactersToSearch(from?.date, minCharactersToSearch) &&
+            isValueBiggerThanMinCharactersToSearch(to?.date, minCharactersToSearch)
         );
     }
 
     return true;
 };
 
-// eslint-disable-next-line complexity
+const shouldSkipMinCharsValidation = (value: any) =>
+    value === undefined || value === null || value === EMPTY_VALUE_FILTER || value === NOT_EMPTY_VALUE_FILTER;
+
 export const isValidMinCharactersToSearch = (value: any, minCharactersToSearch: number) => {
-    if (value === undefined || value === EMPTY_VALUE_FILTER || value === NOT_EMPTY_VALUE_FILTER) {
+    if (shouldSkipMinCharsValidation(value)) {
         return true;
     }